+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- fixed: Disabling pin-login while in duress mode disables pin-login for the main login.
 - fixed: `loginWithPassword` bug disabled duress mode when doing an online login.
 
 ## 2.27.4 (2025-05-13)